docs: enhanced docs
