--- src/components/settings/sections/directory-settings-section.ts
@@ -1,57 +1,0 @@
-import { SettingsUIBuilder } from "@real1ty-obsidian-plugins/utils";
-
-import type { SettingsSection } from "../types";
-import type { NexusPropertiesSettingsSchema } from "src/types/settings";
-
-export class DirectorySettingsSection implements SettingsSection {
-	readonly id = "directories";
-	readonly label = "Directory Scanning";
-
-	constructor(
-		private readonly uiBuilder: SettingsUIBuilder<typeof NexusPropertiesSettingsSchema>
-	) {}
-
-	render(container: HTMLElement): void {
-		this.uiBuilder.addArrayManager(container, {
-			key: "directories",
-			name: "Directory scanning",
-			desc: 'Configure which directories to scan for files with relationships. Use "*" to scan all directories, or specify individual directories to limit scanning.',
-			placeholder: "Directory path (e.g., Projects or Notes/Work)",
-			addButtonText: "Add",
-			removeButtonText: "Remove",
-			emptyArrayFallback: "*",
-			preventEmpty: true,
-			itemDescriptionFn: (item: unknown) => {
-				const dir = String(item);
-				return dir === "*" ? "Scan all directories" : `Includes all subdirectories: ${dir}/**`;
-			},
-			onBeforeAdd: async (newItem: unknown, currentItems: unknown[]) => {
-				const newDir = String(newItem);
-				let newDirs = [...currentItems];
-
-				if (newDir !== "*" && newDirs.includes("*")) {
-					newDirs = newDirs.filter((dir) => dir !== "*");
-				}
-
-				if (!newDirs.includes(newDir)) {
-					newDirs.push(newDir);
-				}
-
-				return newDirs;
-			},
-			onBeforeRemove: async (itemToRemove: unknown, currentItems: unknown[]) => {
-				const newDirs = currentItems.filter((dir) => dir !== itemToRemove);
-				return newDirs.length === 0 ? ["*"] : newDirs;
-			},
-			quickActions: [
-				{
-					name: "Reset to scan all directories",
-					desc: "Clear all specific directories and scan the entire vault",
-					buttonText: "Scan all",
-					condition: (currentItems: unknown[]) => !currentItems.includes("*"),
-					action: async () => ["*"],
-				},
-			],
-		});
-	}
-}
