+++ src/components/settings/settings-tab.ts
@@ -1,20 +1,17 @@
 import { SettingsUIBuilder } from "@real1ty-obsidian-plugins/utils";
-import { type App, PluginSettingTab, Setting } from "obsidian";
+import { type App, PluginSettingTab } from "obsidian";
 
 import type NexusPropertiesPlugin from "src/main";
 import type { NexusPropertiesSettingsSchema } from "src/types/settings";
-
-import type { SettingsSection } from "./types";
-import { ColorSettingsSection } from "./sections/color-settings-section";
 import { DirectRelationshipsSection } from "./sections/direct-relationships-section";
-import { DirectorySettingsSection } from "./sections/directory-settings-section";
+import { DirectoriesSection } from "./sections/directories-section";
 import { ExamplesSection } from "./sections/examples-section";
-import { FilteringSettingsSection } from "./sections/filtering-settings-section";
 import { GraphDisplaySettingsSection } from "./sections/graph-display-section";
-import { IndexingSection } from "./sections/indexing-section";
 import { NodeCreationSettingsSection } from "./sections/node-creation-section";
 import { PropertyDisplaySettingsSection } from "./sections/property-display-section";
+import { RulesSection } from "./sections/rules-section";
 import { UserInterfaceSettingsSection } from "./sections/user-interface-section";
+import type { SettingsSection } from "./types";
 
 const SPONSOR_URL = "https://github.com/sponsors/Real1tyy";
 
@@ -41,19 +38,21 @@ export class NexusPropertiesSettingsTab extends PluginSettingTab {
 		containerEl.createEl("h1", { text: "Nexus Properties Settings" });
 
 		if (this.sections.length > 0) {
-			const sectionSelect = new Setting(containerEl)
-				.setName("Section")
-				.setDesc("Choose which settings section to configure");
+			const navContainer = containerEl.createDiv("nexus-settings-nav");
 
-			sectionSelect.addDropdown((dropdown) => {
-				for (const section of this.sections) {
-					dropdown.addOption(section.id, section.label);
+			this.sections.forEach((section) => {
+				const button = navContainer.createEl("button", {
+					text: section.label,
+					cls: "nexus-settings-nav-button",
+				});
+
+				if (this.selectedSectionId === section.id) {
+					button.addClass("nexus-settings-nav-button-active");
 				}
 
-				dropdown.setValue(this.selectedSectionId);
-				dropdown.onChange((value) => {
-					this.selectedSectionId = value;
-					this.renderSelectedSection();
+				button.addEventListener("click", () => {
+					this.selectedSectionId = section.id;
+					this.display();
 				});
 			});
 		}
@@ -86,10 +85,8 @@ export class NexusPropertiesSettingsTab extends PluginSettingTab {
 			new UserInterfaceSettingsSection(this.uiBuilder),
 			new GraphDisplaySettingsSection(this.uiBuilder),
 			new PropertyDisplaySettingsSection(this.uiBuilder),
-			new ColorSettingsSection(this.plugin),
-			new FilteringSettingsSection(this.plugin, this.uiBuilder),
-			new IndexingSection(this.plugin),
-			new DirectorySettingsSection(this.uiBuilder),
+			new RulesSection(this.plugin, this.uiBuilder),
+			new DirectoriesSection(this.plugin, this.uiBuilder),
 			new DirectRelationshipsSection(this.uiBuilder),
 			new NodeCreationSettingsSection(this.plugin, this.uiBuilder),
 			new ExamplesSection(this.plugin),
