+++ styles.css
@@ -103,6 +103,49 @@
 	line-height: 1.5;
 }
 
+/* Settings navigation tabs */
+.nexus-settings-nav {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 8px;
+	margin: 1.5em 0;
+	padding-bottom: 1em;
+	border-bottom: 2px solid var(--background-modifier-border);
+}
+
+.nexus-settings-nav-button {
+	padding: 8px 16px;
+	background-color: var(--background-secondary);
+	border: 1px solid var(--background-modifier-border);
+	border-radius: 6px;
+	color: var(--text-normal);
+	cursor: pointer;
+	font-size: 0.95em;
+	font-weight: 500;
+	transition: all 0.2s ease;
+}
+
+.nexus-settings-nav-button:hover {
+	background-color: var(--background-modifier-hover);
+	border-color: var(--interactive-accent);
+}
+
+.nexus-settings-nav-button-active {
+	background-color: var(--interactive-accent);
+	color: var(--text-on-accent);
+	border-color: var(--interactive-accent);
+}
+
+.nexus-settings-nav-button-active:hover {
+	background-color: var(--interactive-accent-hover);
+	border-color: var(--interactive-accent-hover);
+}
+
+/* Settings section container */
+.nexus-settings-section-container {
+	margin-top: 1em;
+}
+
 /* Color example list items */
 .color-example-item {
 	display: flex;
